requirements.txt
black
click
isort
ruamel.yaml
pyyaml
pytest
pandas
matplotlib
